Post-season code cleanup - Split up shooter - Destructed teleop positioning command into different files for each command - Changed auto short shoot and etc. to have one command that takes setpoints (much less file clutter) - prefified multiple numbers

+++ src/y2024/include/subsystems/drivetrain.h
@@ -148,7 +148,7 @@ class DrivetrainSubsystem
     frc846::Pref<double> angle_offset_factor_{april_tags_named_, "april_angle_factor", 1.0};
 
     bool aprilFrameRequested=false;
-    long aprilFrameRequest=0;
+    double aprilFrameRequest=0;
     frc846::util::Position poseAtLastRequest;
 
     nt::NetworkTableInstance nt_table =
